[media] Use VideoFrame::Plane typed enum instead of nameless enum
Use VideoFrame::Plane typed enum everywhere instead of previous
nameless enum within VideoFrame. Used tools/git/mffr.py for the
mass renaming.

We'll later use this enum to properly identify plane configs and set
VideoFrame::data

Bug: 40256189
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5530163
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1300035}

+++ media/gpu/ipc/service/picture_buffer_manager.cc
@@ -230,7 +230,7 @@ class PictureBufferManagerImpl : public PictureBufferManager {
 
     // If this |picture| has a SharedImage, then keep a reference to the
     // SharedImage in |picture_buffer_data|.
-    for (int i = 0; i < VideoFrame::kMaxPlanes; i++) {
+    for (size_t i = 0; i < VideoFrame::kMaxPlanes; i++) {
       auto image = picture.scoped_shared_image(i);
       picture_buffer_data.scoped_shared_images[i] = std::move(image);
     }
@@ -255,7 +255,7 @@ class PictureBufferManagerImpl : public PictureBufferManager {
 
       CHECK(picture_buffer_data.scoped_shared_images[0]);
 
-      for (int i = 0; i < VideoFrame::kMaxPlanes; i++) {
+      for (size_t i = 0; i < VideoFrame::kMaxPlanes; i++) {
         const auto& image = picture_buffer_data.scoped_shared_images[i];
         if (image) {
           mailbox_holders[i] = image->GetMailboxHolder();
